Add application files and necessary configurations
This commit introduces several new files and changes. It adds a '.eslintrc.cjs' and '.gitignore' files for project settings, an 'index.css' for styling, and a 'db.js' for the application data. It also adds component files for the core functionalities, including 'App.jsx', 'Header.jsx', and 'Guitar.jsx'. Lastly, it extends eslint configurations to add React specific linting rules.