Add conditional rendering to cart in Header component
Empty cart message "El carrito esta vacio" will now be displayed if the cart is empty. Cart contents (including a table of items) will only be displayed if the cart has one or more items. This improves user experience on the website as irrelevant information is hidden when not needed.