Add functionality to remove items from cart
Adjusted the Header component to take a second prop, 'removeFromCart', for the removal of items from the cart. This addition required modifying instances of 'cartState.cart' in Header.jsx to 'cart' as the component now directly receives the 'cart' and 'removeFromCart' properties. Also, added an 'onClick' event to the remove button in the cart display.