Add quantity control and clear cart functionality
Implemented features to control the quantity of items in the shopping cart by adding increaseQuantity and decrementQuantity functions. Also, a clearCart functionality is added to allow users to empty the cart. The MAX_ITEMS and MIN_ITEMS constants are used to limit the quantity of items per product in the cart.